src/server/requirements.txt
black==24.8.0
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pytest==8.3.2
pytest-flask==1.3.0
python-dotenv==1.0.1
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
